fix(core): Require a whole number for the expression engine memory limit
The memory limit is expressed in megabytes, so a fraction has no
meaningful unit. The isolate tolerates one by truncating it, which
means the effective limit silently differs from the configured value.
Reject it at the config boundary instead, matching the timeout schema.

+++ packages/@n8n/config/src/configs/expression-engine.config.ts
@@ -33,8 +33,10 @@ export class ExpressionEngineConfig {
 	/**
 	 * Memory limit in MB for the V8 isolate used by the VM bridge.
 	 * The isolate requires at least 8MB, and a negative value leaves it without a limit.
+	 * It tolerates a fractional value by truncating it, but whole megabytes are the only
+	 * meaningful unit here, so this rejects a fraction rather than silently truncating.
 	 */
-	@Env('N8N_EXPRESSION_ENGINE_MEMORY_LIMIT', z.number({ coerce: true }).min(8))
+	@Env('N8N_EXPRESSION_ENGINE_MEMORY_LIMIT', z.number({ coerce: true }).int().min(8))
 	bridgeMemoryLimit: number = 128;
 
 	/**
